refactor: add VercelUseChatRuntime
